Move `MockLogger` to a `mocks` directory.
